+++ src/components/ui/DataList.module.css
@@ -1,0 +1,95 @@
+.main_all {
+  padding: 60px 0;
+  min-height: 1000px;
+}
+.right {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+}
+
+.sort_box {
+  position: relative;
+  z-index: 3;
+}
+.sort_box > p {
+  font-size: 1.4rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 105px;
+  height: 33px;
+  gap: 6px;
+  border-radius: 5px;
+  font-weight: 700;
+  background-color: var(--gray-200);
+  cursor: pointer;
+}
+.sort_box > p > svg {
+  margin-top: -2px;
+}
+.sort_box > ul {
+  position: absolute;
+  top: calc(100% + 8px);
+  left: 0;
+  z-index: 2;
+  width: 100%;
+  border-radius: 6px;
+  border: 1px solid var(--gray-200);
+  padding: 4px 0;
+  background-color: #fff;
+}
+.sort_box > ul > li {
+  padding: 8px 0;
+  text-align: center;
+  font-size: 1.4rem;
+  line-height: 1.5714;
+  cursor: pointer;
+}
+.sort_box > ul > li + li {
+  border-top: 1px solid var(--gray-200);
+}
+.right > div > button {
+  font-size: 1.4rem;
+  background-color: var(--red-300);
+  max-width: none;
+  width: auto;
+}
+.filter {
+  position: relative;
+  z-index: 3;
+}
+.filter > div {
+  position: absolute;
+  top: calc(100% + 8px);
+  right: 0;
+  width: 390px;
+  background-color: #fff;
+}
+.page {
+  margin-top: 40px;
+}
+.no_data {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 200px;
+  font-size: 1.8rem;
+}
+
+@media (max-width: 744px) {
+  .main_all {
+    padding: 40px 0;
+  }
+  .page {
+    margin-top: 20px;
+  }
+  .filter > div {
+    position: fixed;
+    top: 0;
+    right: 0;
+    max-width: none;
+    width: 100%;
+    height: 100dvh;
+  }
+}
