+++ src/pages/view/[shopId]/[noticeId]/view.module.css
@@ -11,197 +11,10 @@
   z-index: -1;
   background-color: var(--gray-100);
 }
-.view_content {
-  padding: 60px 0;
-}
-
-.category {
-  font-size: 1.6rem;
-  font-weight: 700;
-  color: var(--primary);
-  margin-bottom: 8px;
-}
-
-.box {
-  display: flex;
-  background-color: var(--white);
-  border-radius: 12px;
-  border: 1px solid var(--gray-200);
-  padding: 24px;
-}
-.box .thumbnail {
-  position: relative;
-  width: 60%;
-  height: 308px;
-  border-radius: 12px;
-  overflow: hidden;
-}
-.box .thumbnail > span {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  display: none;
-  align-items: center;
-  justify-content: center;
-  background-color: rgba(0, 0, 0, 0.6);
-  color: var(--gray-300);
-  font-size: 2.8rem;
-  z-index: 2;
-  font-weight: 700;
-}
-.box.end .thumbnail > span {
-  display: flex;
-}
-.box .thumbnail > img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-.content {
-  width: 40%;
-  display: flex;
-  flex-direction: column;
-  padding-top: 16px;
-  padding-left: 30px;
-}
-.content > p {
-  font-size: 1.6rem;
-  color: var(--primary);
-  font-weight: 700;
-  margin-bottom: 8px;
-}
-.content > .price {
-  display: flex;
-  align-items: center;
-  margin-bottom: 12px;
-}
-.content > .price > p {
-  font-size: 2.8rem;
-  font-weight: 700;
-  letter-spacing: -0.56px;
-}
-.content > ul {
-  margin-bottom: 12px;
-}
-.content > ul > li {
-  display: flex;
-  align-items: center;
-}
-.content > ul > li + li {
-  margin-top: 12px;
-}
-.content > ul > li > svg {
-  width: 20px;
-  height: 20px;
-  margin-right: 8px;
-}
-.content > ul > li > span {
-  font-size: 1.6rem;
-  line-height: 1.625;
-  color: var(--gray-500);
-}
-.content .txt {
-  font-size: 1.6rem;
-  line-height: 1.625;
-  margin-bottom: 14px;
-}
-.content .txt > p {
-  white-space: pre-line;
-}
-.content > button {
-  max-width: none;
-  width: 100%;
-  margin-top: auto;
-  font-size: 1.6rem;
-}
-.description {
-  margin-top: 24px;
-  background-color: var(--gray-200);
-  padding: 32px;
-  border-radius: 12px;
-}
-.description > h3 {
-  font-size: 1.6rem;
-  font-weight: 700;
-  margin-bottom: 12px;
-}
-.description > p {
-  font-size: 1.6rem;
-  line-height: 1.625;
-  white-space: pre-line;
-}
 .view_recent {
   padding: 60px 0;
 }
-
-@media (max-width: 964px) {
-  .box {
-    display: block;
-  }
-  .box .thumbnail {
-    position: relative;
-    width: 100%;
-    height: 0;
-    padding-top: 57.1202%;
-    border-radius: 12px;
-    overflow: hidden;
-  }
-  .box .thumbnail > img {
-    position: absolute;
-    top: 0;
-    left: 0;
-  }
-  .content {
-    width: 100%;
-    padding-left: 0;
-  }
-}
 @media (max-width: 744px) {
-  .box {
-    padding: 20px;
-  }
-  .box .thumbnail > span {
-    font-size: 2rem;
-  }
-  .view_content {
-    padding: 40px 0;
-  }
-  .content > .price > p {
-    font-size: 2.4rem;
-  }
-  .content > ul > li + li {
-    margin-top: 8px;
-  }
-  .content > ul > li > svg {
-    width: 16px;
-    height: 16px;
-    margin-right: 6px;
-  }
-  .content > ul > li > span {
-    font-size: 1.4rem;
-    line-height: 1.5714;
-  }
-  .content .txt {
-    font-size: 1.4rem;
-    line-height: 1.5714;
-    margin-bottom: 24px;
-  }
-  .content > button {
-    font-size: 1.4rem;
-  }
-  .description {
-    margin-top: 12px;
-    padding: 20px;
-  }
-  .description > h3 {
-    font-size: 1.4rem;
-    margin-bottom: 8px;
-  }
-  .description > p {
-    font-size: 1.4rem;
-  }
   .view_recent {
     padding: 40px 0;
   }
